Move crypto implementations to src/crypto/
Conflicts:
	src/Makefile.am
