+++ src/Makefile.am
@@ -79,8 +79,9 @@ BITCOIN_CORE_H = \
   script.h \
   scrypt.h \
   serialize.h \
-  sha2.h \
-  sha1.h \
+  crypto/scrypt.h \
+  crypto/sha2.h \
+  crypto/sha1.h \
   sync.h \
   threadsafety.h \
   tinyformat.h \
@@ -163,9 +164,9 @@ libbitcoin_common_a_SOURCES = \
   protocol.cpp \
   rpcprotocol.cpp \
   script.cpp \
-  scrypt.cpp \
-  sha1.cpp \
-  sha2.cpp \
+  crypto/scrypt.cpp \
+  crypto/sha1.cpp \
+  crypto/sha2.cpp \
   sync.cpp \
   util.cpp \
   version.cpp \
