+++ src/crypto/scrypt-sse2.cpp
