+++ src/crypto/scrypt.h
