+++ src/crypto/sha2.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "sha2.h"
+#include "crypto/sha2.h"
 
 #include <string.h>
