+++ src/main.h
@@ -18,7 +18,7 @@
 #include "core.h"
 #include "net.h"
 #include "script.h"
-#include "scrypt.h"
+#include "crypto/scrypt.h"
 #include "sync.h"
 #include "txmempool.h"
 #include "uint256.h"
