+++ src/script.cpp
@@ -9,8 +9,8 @@
 #include "hash.h"
 #include "key.h"
 #include "keystore.h"
-#include "sha1.h"
-#include "sha2.h"
+#include "crypto/sha1.h"
+#include "crypto/sha2.h"
 #include "sync.h"
 #include "uint256.h"
 #include "util.h"
