+++ src/test/sha1_tests.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "sha1.h"
+#include "crypto/sha1.h"
 #include "util.h"
 
 #include <vector>
